[native] Get rid of nav-reducer.js
