+++ native/navigation/root-navigator.react.js
@@ -15,6 +15,7 @@ import {
   type NavigationStackTransitionProps,
 } from 'react-navigation-stack';
 import { Keyboard } from 'react-native';
+import { useScreens } from 'react-native-screens';
 
 import {
   LoggedOutModalRouteName,
@@ -36,6 +37,9 @@ import CustomServerModal from '../more/custom-server-modal.react';
 import ColorPickerModal from '../chat/settings/color-picker-modal.react';
 import ComposeSubthreadModal from '../chat/settings/compose-subthread-modal.react';
 import RootRouter from './root-router';
+
+// eslint-disable-next-line react-hooks/rules-of-hooks
+useScreens();
 
 type NavigationProp = NavigationStackProp<NavigationState> & {
   logIn: () => void,
